+++ src/features/marketplace/components/rewards/ItemDetail.tsx
@@ -1,0 +1,10 @@
+import { CloseButtonPanel } from "features/game/components/CloseablePanel";
+import React from "react";
+
+export const ItemDetail: React.FC = () => {
+  return (
+    <CloseButtonPanel>
+      <div></div>
+    </CloseButtonPanel>
+  );
+};
